refactor(web): split domain lib utilities into per-feature lib folders
calendarUtils/freeSlots -> features/calendar/lib, taskUtils ->
features/tasks/lib, subscriptionUtils -> features/subscription/lib.
constants.ts (lobby-type styling, task-status/priority labels, and
the QUERY_KEYS cache-key registry) is split the same way, with each
feature keeping its own QUERY_KEYS object so call sites didn't need
renaming, only their import path. src/lib/ now holds only the truly
cross-feature infra: apiClient, apiErrors, and the cn() class-merge
helper.
