+++ lined-web/src/features/notifications/lib/constants.ts
@@ -1,0 +1,6 @@
+export const QUERY_KEYS = {
+  notificationPreferences: ['notifications', 'preferences'] as const,
+  lobbyNotificationPreferences: (lobbyId: number) =>
+    ['notifications', 'lobby-preferences', lobbyId] as const,
+  myNotifications: ['notifications', 'mine'] as const,
+} as const;
