+++ lined-web/src/features/subscription/lib/constants.ts
@@ -1,0 +1,5 @@
+export const QUERY_KEYS = {
+  plans: ['plans'] as const,
+  activeSubscription: (userId: number) => ['subscriptions', userId, 'active'] as const,
+  subscriptionHistory: (userId: number) => ['subscriptions', userId, 'history'] as const,
+} as const;
